Minor plot fixes

+++ settings.py
@@ -47,8 +47,8 @@ class Run(BaseSettings):
 class Plot(BaseSettings):
     show: bool = False
     n_queries: int = 7
-    limit_with_io: int = 20
-    limit_without_io: int = 15
+    limit_with_io: float = 20.0
+    limit_without_io: float = 15.0
 
     model_config = SettingsConfigDict(
         env_prefix="plot_", env_file=".env", extra="ignore"
